The player can now be damaged in battles. There isn't really any strategy or chance involved yet.

+++ src/item/mod.rs
@@ -1,6 +1,6 @@
 pub mod equipment {
     pub struct Weapon {
-        pub strength: usize,
+        pub strength: u32,
     }
 }
